fix: updated holder dialog
