+++ apps/website/components/templates/gate-view/gate-view.tsx
@@ -1,6 +1,3 @@
-import dynamic from 'next/dynamic';
-import { useState, ComponentType } from 'react';
-
 import { PartialDeep } from 'type-fest';
 
 import { Grid, Divider } from '@mui/material';
@@ -8,21 +5,14 @@ import { Grid, Divider } from '@mui/material';
 import { useCredentialByGateId } from '../../../hooks/use-credential-by-gate-id';
 import { useGateStatus } from '../../../hooks/use-gate-status';
 import { Gates } from '../../../services/hasura/types';
-import type { Props as HolderDialogProps } from '../../organisms/holder-dialog';
 import { GateViewSidebar } from './gate-view-sidebar';
 import { GateViewTasks } from './gate-view-tasks';
-
-const HolderDialog: ComponentType<HolderDialogProps> = dynamic(
-  () => import('../../organisms/holder-dialog').then((mod) => mod.HolderDialog),
-  { ssr: false }
-);
 
 type GateViewProps = {
   gateProps: PartialDeep<Gates>;
 };
 
 export function GateViewTemplate({ gateProps }: GateViewProps) {
-  const [isHolderDialog, setIsHolderDialog] = useState(false);
   const gateStatus = useGateStatus(gateProps);
   const credential = useCredentialByGateId({ gateId: gateProps?.id });
 
@@ -34,13 +24,6 @@ export function GateViewTemplate({ gateProps }: GateViewProps) {
         flexDirection: { xs: 'column', md: 'row' },
       }}
     >
-      <HolderDialog
-        {...{
-          isHolderDialog,
-          setIsHolderDialog,
-          credentialId: gateProps?.id,
-        }}
-      />
       <GateViewSidebar
         completedGate={gateStatus.isCompleted}
         credential={credential}
